adding auth for supplier

+++ supplier/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
 from .models import Supplier
 from django.contrib import messages
+from propensi.utils import is_gudang_pusat, is_manager, is_staff
 
 # Create your views here.
 def create_supplier(request):
+    if(auth_gedung_pusat(request) == False):
+        return render(request, 'error/403.html')
     if(request.POST):
         supplier = Supplier.objects.create()
         supplier.nama = request.POST["nama"]
@@ -16,11 +19,16 @@ def create_supplier(request):
     return render(request, 'list_supplier_datatables.html')
 
 def list_supplier(request):
+    print(auth_gedung_pusat(request))
+    if(auth_gedung_pusat(request) == False):
+        return render(request, 'error/403.html')
     suppliers = Supplier.objects.all()
     response = {'suppliers':suppliers}
     return render(request, 'list_supplier_datatables.html', response)
 
 def update_supplier(request,id):
+    if(auth_gedung_pusat(request) == False):
+        return render(request, 'error/403.html')
     supplier = Supplier.objects.get(id_supplier=id)
     if(request.POST):
         supplier.nama = request.POST["nama"]
@@ -33,10 +41,21 @@ def update_supplier(request,id):
     return render(request, 'list_supplier_datatables.html', {'supplier':supplier})
 
 def get_supplier(request,id):
-        supplier = Supplier.objects.get(id_supplier=id)  
-        return render(request,'list_supplier_datatables.html', {'supplier':supplier})
+    if(auth_gedung_pusat(request) == False):
+        return render(request, 'error/403.html')
+    supplier = Supplier.objects.get(id_supplier=id)  
+    return render(request,'list_supplier_datatables.html', {'supplier':supplier})
 
 def delete_supplier(request,id):
+    if(auth_gedung_pusat(request) == False):
+        return render(request, 'error/403.html')
     supplier = Supplier.objects.get(id_supplier=id)
     supplier.delete()
     return redirect("/supplier/list_supplier")
+
+def auth_gedung_pusat(request):
+    if (is_manager(request) | is_staff(request)):
+        if (is_gudang_pusat(request)):
+            return True
+        return False
+    return False
